+++ lib/screens/settings/color_settings.dart
@@ -5,6 +5,7 @@
 
 import '../../utils/export.dart';
 import '../../widgets/export.dart';
+import 'package:efui_bios/efui_bios.dart';
 
 import 'package:flutter/material.dart';
 import 'package:empathetech_flutter_ui/empathetech_flutter_ui.dart';
@@ -36,6 +37,19 @@ class ColorSettingsScreen extends StatelessWidget {
           ),
           useImageDecoration: false,
         ),
-        fab: EzBackFAB(context),
+        fabs: <Widget>[
+          ezSpacer,
+          EzConfigFAB(
+            context,
+            appName: appName,
+            androidPackage: packageName,
+          ),
+          ezSpacer,
+          const EzBackFAB(
+            hold4Feedback: true,
+            appName: appName,
+            supportEmail: empathSupport,
+          ),
+        ],
       );
 }