+++ lib/screens/settings/text_settings.dart
@@ -3,7 +3,9 @@
  * See LICENSE for distribution and usage details.
  */
 
+import '../../utils/export.dart';
 import '../../widgets/export.dart';
+import 'package:efui_bios/efui_bios.dart';
 
 import 'package:flutter/material.dart';
 import 'package:empathetech_flutter_ui/empathetech_flutter_ui.dart';
@@ -16,6 +18,19 @@ class TextSettingsScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) => SosScaffold(
         EzScreen(EzTextSettings(target: target), useImageDecoration: false),
-        fab: EzBackFAB(context),
+        fabs: <Widget>[
+          ezSpacer,
+          EzConfigFAB(
+            context,
+            appName: appName,
+            androidPackage: packageName,
+          ),
+          ezSpacer,
+          const EzBackFAB(
+            hold4Feedback: true,
+            appName: appName,
+            supportEmail: empathSupport,
+          ),
+        ],
       );
 }