+++ lib/utils/consts.dart
@@ -7,63 +7,97 @@ import 'export.dart';
 
 import 'package:empathetech_flutter_ui/empathetech_flutter_ui.dart';
 
-/// 'InstaSOS'
+// App config //
+
+/// InstaSOS
 const String appName = 'InstaSOS';
 
 /// net.empathetech.sos
 const String packageName = 'net.empathetech.sos';
 
-/// 'taskRunning'
-const String taskRunningKey = 'taskRunning';
+// BTS settings //
 
-/// 'showTutorial'
+/// showTutorial
 const String tutorialKey = 'showTutorial';
 
-/// 'Video'
-const String videoColorKey = 'Video';
+/// showContactsMsg
+const String partialContactsKey = 'showContactsMsg';
 
-/// 'emc'
-const String emcKey = 'emc';
-
-/// 'linkType'
-const String linkTypeKey = 'linkType';
-
-/// 'savedTab'
+/// savedTab
 const String savedTabKey = 'savedTab';
 
-/// 'walking'
+/// walking
 const String walkingTab = 'walking';
 
-/// 'driving'
+/// driving
 const String drivingTab = 'driving';
 
-/// 'atHome'
+/// atHome
 const String atHomeTab = 'atHome';
 
-/// 'notifyOnOpen'
+/// taskRunning
+const String taskRunningKey = 'taskRunning';
+
+// SOS settings //
+
+/// notifyOnOpen
 const String onOpenKey = 'sosOnOpen';
 
-/// 'notifyOnClose'
+/// notifyOnClose
 const String onCloseKey = 'sosOnClose';
 
-/// 'sosOnInterrupt'
+/// sosOnInterrupt
 const String onInterruptKey = 'sosOnInterrupt';
 
-/// 'autoShareMedia'
+/// autoShareMedia
 const String autoShareKey = 'autoShareMedia';
 
+/// emc
+const String emcKey = 'emc';
+
+/// linkType
+const String linkTypeKey = 'linkType';
+
+// Color settings //
+
+/// Video
+const String videoColorKey = 'Video';
+
+// EzConfig default //
+
+/// defaults for [EzConfig.init]
 final Map<String, Object> sosConfig = <String, Object>{
+  // EFUI
   ...empathMobileConfig,
   darkTextBackgroundOpacityKey: 0.8,
   lightTextBackgroundOpacityKey: 0.8,
-  videoColorKey: 0xFFFF0000,
-  emcKey: <String>[],
-  linkTypeKey: LinkType.google.name,
-  savedTabKey: walkingTab,
+
+  // BTS
   tutorialKey: true,
+  partialContactsKey: true,
+  savedTabKey: walkingTab,
+  taskRunningKey: false,
+
+  // SOS
   onOpenKey: false,
   onCloseKey: false,
   onInterruptKey: true,
   autoShareKey: true,
-  taskRunningKey: false,
+  emcKey: <String>[],
+  linkTypeKey: LinkType.google.name,
+
+  // Color
+  videoColorKey: 0xFFFF0000,
+};
+
+// Local assets //
+
+/// assets/images/og-baddie.jpg
+const String ladyLiberty = 'assets/images/og-baddie.jpg';
+
+/// [ladyLiberty]
+const Set<String> assetPaths = <String>{ladyLiberty};
+
+const Map<String, String> credits = <String, String>{
+  ladyLiberty: 'Public domain; CC0 1.0 Universal',
 };