+++ lib/widgets/emc.dart
@@ -5,49 +5,9 @@
 
 import '../utils/export.dart';
 
-import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:empathetech_flutter_ui/empathetech_flutter_ui.dart';
 import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
-
-class ContactTile extends StatelessWidget {
-  final String number;
-  final bool enabled;
-  final VoidCallback onRemove;
-
-  const ContactTile({
-    super.key,
-    required this.number,
-    required this.enabled,
-    required this.onRemove,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final Lang l10n = Lang.of(context)!;
-    final EFUILang el10n = ezL10n(context);
-
-    return EzScrollView(
-      mainAxisSize: MainAxisSize.min,
-      scrollDirection: Axis.horizontal,
-      reverseHands: true,
-      children: <Widget>[
-        Text(number, style: Theme.of(context).textTheme.bodyLarge),
-        const EzSpacer(vertical: false),
-        EzIconButton(
-          icon: Icon(
-            PlatformIcons(context).removeCircledOutline,
-            semanticLabel:
-                '${l10n.ssRemoveHint}.${enabled ? '' : ' ${el10n.gDisabled}.'}',
-          ),
-          enabled: enabled,
-          onPressed: onRemove,
-          tooltip: l10n.ssRemoveHint,
-        )
-      ],
-    );
-  }
-}
 
 class ContactList extends StatefulWidget {
   const ContactList({super.key});
@@ -61,96 +21,177 @@ class _ContactListState extends State<ContactList> {
 
   final double margin = EzConfig.get(marginKey);
   final double padding = EzConfig.get(paddingKey);
-  final double spacing = EzConfig.get(paddingKey);
-
+  final double spacing = EzConfig.get(spacingKey);
   final double iconSize = EzConfig.get(iconSizeKey);
 
+  late final EzSpacer listSpacer = EzSpacer(space: spacing - margin * 2);
+
   late final Lang l10n = Lang.of(context)!;
-
-  // Define build data //
-
-  List<String> emc = EzConfig.get(emcKey);
-  late int heightMod = min(5, emc.length);
-
-  late final Size numSize = ezTextSize(
-    '+44 (444) 444-4444',
-    context: context,
-    style: Theme.of(context).textTheme.bodyLarge,
-  );
-
-  late final double listWidth = 2 * margin +
-      max(iconSize + padding, kMinInteractiveDimension) +
-      numSize.width +
-      spacing;
-
-  // Define custom functions //
-
-  double listHeight() => max(
-      margin +
-          (padding + numSize.height) * heightMod +
-          spacing * (heightMod - 1),
-      2 * margin + kMinInteractiveDimension);
+  late final EFUILang el10n = ezL10n(context);
 
   // Return the build //
+
+  List<String> emc = EzConfig.get(emcKey);
 
   @override
   Widget build(BuildContext context) {
     final TextTheme textTheme = Theme.of(context).textTheme;
+    final ColorScheme colorScheme = Theme.of(context).colorScheme;
 
     return Column(
       mainAxisSize: MainAxisSize.min,
       children: <Widget>[
         // Title && add button
         EzScrollView(
-          scrollDirection: Axis.horizontal,
           reverseHands: true,
           mainAxisSize: MainAxisSize.min,
+          scrollDirection: Axis.horizontal,
           children: <Widget>[
             Text(l10n.ssEMC, style: textTheme.titleLarge),
-            EzMargin(),
+            ezRowMargin,
             EzIconButton(
               icon: Icon(
                 PlatformIcons(context).addCircledOutline,
                 semanticLabel: l10n.ssAddHint,
               ),
               onPressed: () async {
-                emc = await addEMC(context, emc, loop: false) ?? emc;
-                setState(() => heightMod = min(5, emc.length));
+                emc = await addEMC(context, emc,
+                        loop: false, l10n: l10n, el10n: el10n) ??
+                    emc;
+                setState(() {});
               },
               tooltip: l10n.ssAddHint,
             ),
           ],
         ),
-        EzMargin(),
+        ezMargin,
 
         // List of numbers (with remove buttons)
         ConstrainedBox(
-          constraints: BoxConstraints(
-            maxHeight: listHeight(),
-            maxWidth: listWidth,
-          ),
+          constraints: BoxConstraints(maxHeight: heightOf(context) / 3),
           child: Card(
-            child: Center(
-              child: EzScrollView(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: emc
-                    .map((String number) => ContactTile(
-                          key: ValueKey<String>(number),
-                          number: number,
-                          enabled: emc.length > 1,
-                          onRemove: () async {
-                            emc.remove(number);
-                            heightMod = min(5, emc.length);
-                            await EzConfig.setStringList(emcKey, emc);
-                            setState(() {});
-                          },
-                        ))
-                    .toList(),
-              ),
+            child: EzScrollView(
+              mainAxisSize: MainAxisSize.min,
+              children: emc.fold<List<Widget>>(
+                <Widget>[],
+                (List<Widget> acc, String contact) {
+                  final List<String> parts = contact.split(contactSplit);
+                  late final String? initials;
+                  late final String number;
+
+                  if (parts.length == 2) {
+                    initials = parts.first;
+                    number = parts.last;
+                  } else {
+                    initials = null;
+                    number = contact;
+                  }
+
+                  acc.add(_ContactTile(
+                    key: ValueKey<String>(contact),
+                    initials: initials,
+                    number: number,
+                    enabled: emc.length > 1,
+                    onRemove: () async {
+                      emc.remove(contact);
+                      await EzConfig.setStringList(emcKey, emc);
+                      setState(() {});
+                    },
+                    l10n: l10n,
+                    el10n: el10n,
+                    textTheme: textTheme,
+                    colorScheme: colorScheme,
+                    margin: margin,
+                    padding: padding,
+                    iconSize: iconSize,
+                  ));
+                  acc.add(listSpacer);
+                  return acc;
+                },
+              ).sublist(0, emc.length * 2 - 1), // Remove trailing spacer
             ),
           ),
         ),
       ],
     );
   }
+}
+
+class _ContactTile extends StatelessWidget {
+  final String? initials;
+  final String number;
+  final bool enabled;
+  final VoidCallback onRemove;
+  final Lang l10n;
+  final EFUILang el10n;
+  final TextTheme textTheme;
+  final ColorScheme colorScheme;
+  final double margin;
+  final double padding;
+  final double iconSize;
+
+  const _ContactTile({
+    super.key,
+    // Core parameters
+    required this.initials,
+    required this.number,
+    required this.enabled,
+    required this.onRemove,
+
+    // Theme parameters
+    required this.l10n,
+    required this.el10n,
+    required this.textTheme,
+    required this.colorScheme,
+    required this.margin,
+    required this.padding,
+    required this.iconSize,
+  });
+
+  @override
+  Widget build(BuildContext context) => Padding(
+        padding: EdgeInsets.all(margin),
+        child: EzScrollView(
+          reverseHands: true,
+          showScrollHint: true,
+          mainAxisSize: MainAxisSize.min,
+          scrollDirection: Axis.horizontal,
+          children: <Widget>[
+            // Initials coin (if available)
+            if (initials != null) ...<Widget>[
+              CircleAvatar(
+                radius: padding + iconSize / 2,
+                foregroundColor: colorScheme.onSurface,
+                backgroundColor: colorScheme.surfaceContainer,
+                child: Text(
+                  initials!,
+                  style: textTheme.bodyLarge,
+                  textAlign: TextAlign.start,
+                ),
+              ),
+              ezRowMargin,
+            ],
+
+            // Number
+            Text(
+              number,
+              style: textTheme.bodyLarge,
+              textAlign: TextAlign.start,
+            ),
+            ezRowSpacer,
+
+            // Remove button
+            EzIconButton(
+              icon: Icon(
+                PlatformIcons(context).removeCircledOutline,
+                semanticLabel:
+                    '${l10n.ssRemoveHint}.${enabled ? '' : ' ${el10n.gDisabled}.'}',
+              ),
+              enabled: enabled,
+              onPressed: onRemove,
+              tooltip: l10n.ssRemoveHint,
+            ),
+          ],
+        ),
+      );
 }